pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "civicsignal"
version = "0.1.0"
description = "A tool to help you get signal out of civics"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Rashi Abramson"},
    {name = "Clay Rosenthal"}
]
license = {text = "MIT"}
keywords = ["civics", "san-francisco", "government", "data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.15",
    "deepgram-sdk>=4.8.1",
    "feedparser>=6.0.11",
    "openai>=1.98.0",
    "sentence-transformers>=5.0.0",
    "transformers>=4.54.1",
    "requests>=2.32.4",
    "cerebras-cloud-sdk>=1.35.0",
    "click>=8.2.1",
]

[project.urls]
Homepage = "https://github.com/roshkins/CivicSignal"
Repository = "https://github.com/roshkins/CivicSignal"
Issues = "https://github.com/roshkins/CivicSignal/issues"

[project.scripts]
civicsignal = "civicsignal.cli:cli"

[dependency-groups]
dev = [
    "marimo>=0.14.16",
]

[tool.hatch.build.targets.wheel]
packages = ["civicsignal"]
